fixed namespaces & moved scripts to new folder
