--- sdkproject/Assets/MapboxAR/Scripts/FollowTargetTransform.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class FollowTargetTransform : MonoBehaviour
-{
-	[SerializeField]
-	Transform _targetTransform;
-
-	void Update()
-	{
-		transform.position = new Vector3(_targetTransform.position.x, transform.position.y, _targetTransform.position.z);
-	}
-}
