+++ sdkproject/Assets/MapboxAR/Scripts/ManualSync.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 6c7f7345a2ce049d9b39d8ee2f388854
+guid: a4827ae45dfdf430087f9bc2b6dc7a89
 folderAsset: yes
-timeCreated: 1522933685
+timeCreated: 1522933722
 licenseType: Pro
 DefaultImporter:
   externalObjects: {}
