+++ sdkproject/Assets/MapboxAR/Scripts/ManualSync/FollowTargetTransform.cs.meta
