+++ sdkproject/Assets/MapboxAR/Scripts/ManualSync/RotateOnYTargetTransform.cs.meta
