+++ sdkproject/Assets/MapboxAR/Scripts/PlayerSizeFromLocationAccuracy.cs
@@ -1,47 +1,47 @@
-﻿using Mapbox.Unity.Location;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PlayerSizeFromLocationAccuracy : MonoBehaviour
+﻿namespace Mapbox.Examples
 {
+	using Mapbox.Unity.Location;
+	using UnityEngine;
 
-
-	private ILocationProvider _locationProvider;
-	private Vector3 _playerScale = new Vector3(2f, 2f, 2f);
-
-
-	void Start()
+	public class PlayerSizeFromLocationAccuracy : MonoBehaviour
 	{
-		_locationProvider = LocationProviderFactory.Instance.DefaultLocationProvider;
-		if (_locationProvider != null)
+
+		private ILocationProvider _locationProvider;
+		private Vector3 _playerScale = new Vector3(2f, 2f, 2f);
+
+
+		void Start()
 		{
-			_locationProvider.OnLocationUpdated += LocationProvider_OnLocationUpdated;
+			_locationProvider = LocationProviderFactory.Instance.DefaultLocationProvider;
+			if (_locationProvider != null)
+			{
+				_locationProvider.OnLocationUpdated += LocationProvider_OnLocationUpdated;
+			}
 		}
-	}
 
 
-	void OnDestroy()
-	{
-		if (_locationProvider != null)
+		void OnDestroy()
 		{
-			_locationProvider.OnLocationUpdated -= LocationProvider_OnLocationUpdated;
+			if (_locationProvider != null)
+			{
+				_locationProvider.OnLocationUpdated -= LocationProvider_OnLocationUpdated;
+			}
 		}
-	}
 
 
-	void LocationProvider_OnLocationUpdated(Location location)
-	{
-		if (location.Accuracy != 0)
+		void LocationProvider_OnLocationUpdated(Location location)
 		{
-			float halfAcc = location.Accuracy / 2f;
-			_playerScale = new Vector3(halfAcc, halfAcc, halfAcc);
+			if (location.Accuracy != 0)
+			{
+				float halfAcc = location.Accuracy / 2f;
+				_playerScale = new Vector3(halfAcc, halfAcc, halfAcc);
+			}
 		}
-	}
 
 
-	void Update()
-	{
-		transform.localScale = _playerScale;
+		void Update()
+		{
+			transform.localScale = _playerScale;
+		}
 	}
 }
